Refactor election package
